+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Fixed
+ - Fix duplicate Steam application entries
+
 ## [1.6.0] - 2021-08-08
 ### Added
  - Add `protontricks-launch` script to launch Windows executables using Proton app specific Wine prefixes